Add a subdirectory to store ansible for osbs image builds
The Jenkins slave image needs additional dependencies in
order to run cekit osbs builds for rhba images. The ansible
role and data that sets this up references redhat internal
systems and configurations, and so it is stored here rather
than in https://github.com/kiegroup/kie-jenkins-scripts where
the slave image build is maintained. That repo has a script
to clone this repo and install the files appropriately.